corrected div(sigma)
since the Material is dependent on the location this has to be accounted for in the divergence of the stress tensor.
In this step, where nabla is applied to sigma, the material parameters cannot be factored out by assuming them to be constant.
Nabla also has to be applied to lambda and mu. This results in additional terms. the existing terms do not need to be changed!